# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Model Atlas comparison catalog API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface CapabilityRating {
  category: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  score: number;
  /** @nullable */
  notes?: string | null;
}

export interface BenchmarkScore {
  benchmark: string;
  score: number;
  scoreUnit: string;
  sourceUrl: string;
  measuredAt: string;
}

export interface PricingTier {
  tierName: string;
  /** @nullable */
  inputPricePerM: number | null;
  /** @nullable */
  outputPricePerM: number | null;
  effectiveDate: string;
  sourceUrl: string;
}

export interface Provider {
  id: string;
  name: string;
  website: string;
  logoMark: string;
}

export interface ModelSummary {
  id: string;
  slug: string;
  name: string;
  provider: Provider;
  family: string;
  modalities: string[];
  contextWindow: number;
  openWeights: boolean;
  license: string;
  capabilityScore: number;
  inputPricePerM: number;
  outputPricePerM: number;
  complianceCerts: string[];
  deploymentOptions: string[];
  lastVerifiedAt: string;
}

export type ModelProfile = ModelSummary & ({
  maxOutputTokens: number;
  regionsAvailable: string[];
  /** @nullable */
  dataRetentionDays: number | null;
  trainingOptOut: boolean;
  zeroDataRetention: boolean;
  safetyPolicyUrl: string;
  safetyNotes: string;
  sdkLanguages: string[];
  functionCalling: boolean;
  fineTuningSupport: boolean;
  sourceUrl: string;
  capabilities: CapabilityRating[];
  benchmarks: BenchmarkScore[];
  pricing: PricingTier[];
});

export type CatalogSummaryCapabilityLeadersItem = {
  category: string;
  model: string;
  score: number;
};

export interface CatalogSummary {
  modelCount: number;
  providerCount: number;
  benchmarkCount: number;
  lastUpdated: string;
  medianVerificationAgeDays: number;
  priceTrend: string;
  capabilityLeaders: CatalogSummaryCapabilityLeadersItem[];
}

export interface GlossaryTerm {
  term: string;
  shortDefinition: string;
  detail: string;
  sourceUrl: string;
}

export type ChangelogEntryType = typeof ChangelogEntryType[keyof typeof ChangelogEntryType];


export const ChangelogEntryType = {
  pricing: 'pricing',
  benchmark: 'benchmark',
  model: 'model',
  methodology: 'methodology',
} as const;

export interface ChangelogEntry {
  id: string;
  date: string;
  type: ChangelogEntryType;
  title: string;
  summary: string;
  affectedModels: string[];
  sourceUrl: string;
}

export interface FeedbackInput {
  modelSlug: string;
  field: string;
  /** @minLength 5 */
  note: string;
}

export interface FeedbackReceipt {
  id: string;
  message: string;
}

export type ListModelsParams = {
provider?: string;
modality?: string;
openWeights?: boolean;
maxInputPrice?: number;
minContext?: number;
deployment?: string;
search?: string;
sort?: ListModelsSort;
};

export type ListModelsSort = typeof ListModelsSort[keyof typeof ListModelsSort];


export const ListModelsSort = {
  name: 'name',
  price: 'price',
  capability: 'capability',
  context: 'context',
} as const;

export type CompareModelsParams = {
/**
 * Comma-separated model slugs
 */
ids: string;
};

export type ListChangelogParams = {
limit?: number;
};