# lib/api-zod/src/generated/types/benchmarkScore.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Model Atlas comparison catalog API
 * OpenAPI spec version: 0.1.0
 */

export interface BenchmarkScore {
  benchmark: string;
  score: number;
  scoreUnit: string;
  sourceUrl: string;
  measuredAt: Date;
}