# lib/api-zod/src/generated/types/capabilityRating.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Model Atlas comparison catalog API
 * OpenAPI spec version: 0.1.0
 */

export interface CapabilityRating {
  category: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  score: number;
  /** @nullable */
  notes?: string | null;
}