# lib/api-zod/src/generated/types/catalogSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Model Atlas comparison catalog API
 * OpenAPI spec version: 0.1.0
 */
import type { CatalogSummaryCapabilityLeadersItem } from './catalogSummaryCapabilityLeadersItem';

export interface CatalogSummary {
  modelCount: number;
  providerCount: number;
  benchmarkCount: number;
  lastUpdated: Date;
  medianVerificationAgeDays: number;
  priceTrend: string;
  capabilityLeaders: CatalogSummaryCapabilityLeadersItem[];
}