# lib/api-zod/src/generated/types/changelogEntry.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Model Atlas comparison catalog API
 * OpenAPI spec version: 0.1.0
 */
import type { ChangelogEntryType } from './changelogEntryType';

export interface ChangelogEntry {
  id: string;
  date: Date;
  type: ChangelogEntryType;
  title: string;
  summary: string;
  affectedModels: string[];
  sourceUrl: string;
}