# lib/api-zod/src/generated/types/feedbackInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Model Atlas comparison catalog API
 * OpenAPI spec version: 0.1.0
 */

export interface FeedbackInput {
  modelSlug: string;
  field: string;
  /** @minLength 5 */
  note: string;
}