# lib/api-zod/src/generated/types/glossaryTerm.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Model Atlas comparison catalog API
 * OpenAPI spec version: 0.1.0
 */

export interface GlossaryTerm {
  term: string;
  shortDefinition: string;
  detail: string;
  sourceUrl: string;
}