# lib/api-zod/src/generated/types/modelProfile.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Model Atlas comparison catalog API
 * OpenAPI spec version: 0.1.0
 */
import type { BenchmarkScore } from './benchmarkScore';
import type { CapabilityRating } from './capabilityRating';
import type { ModelSummary } from './modelSummary';
import type { PricingTier } from './pricingTier';

export type ModelProfile = ModelSummary & ({
  maxOutputTokens: number;
  regionsAvailable: string[];
  /** @nullable */
  dataRetentionDays: number | null;
  trainingOptOut: boolean;
  zeroDataRetention: boolean;
  safetyPolicyUrl: string;
  safetyNotes: string;
  sdkLanguages: string[];
  functionCalling: boolean;
  fineTuningSupport: boolean;
  sourceUrl: string;
  capabilities: CapabilityRating[];
  benchmarks: BenchmarkScore[];
  pricing: PricingTier[];
});