# lib/api-zod/src/generated/types/modelSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Model Atlas comparison catalog API
 * OpenAPI spec version: 0.1.0
 */
import type { Provider } from './provider';

export interface ModelSummary {
  id: string;
  slug: string;
  name: string;
  provider: Provider;
  family: string;
  modalities: string[];
  contextWindow: number;
  openWeights: boolean;
  license: string;
  capabilityScore: number;
  inputPricePerM: number;
  outputPricePerM: number;
  complianceCerts: string[];
  deploymentOptions: string[];
  lastVerifiedAt: Date;
}