# lib/api-zod/src/generated/types/pricingTier.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Model Atlas comparison catalog API
 * OpenAPI spec version: 0.1.0
 */

export interface PricingTier {
  tierName: string;
  /** @nullable */
  inputPricePerM: number | null;
  /** @nullable */
  outputPricePerM: number | null;
  effectiveDate: Date;
  sourceUrl: string;
}